- Changed scoring of motifs. Scoring now uses predictive evaluation instead of previous ad-hoc method. In short, when descending into a new motif, the likelihood of the discarded positions under the new motif Beta should be low.
- Minimum motif score value changed, as scores are now in a different range than before under the new scoring method.
- Fixed minor issues
- Tweaks to motif merging
- Motif expansion restricted to insertion of explicit bases (A, T, G, C) instead degenerate bases

+++ nanomotif/find_motifs.py
@@ -213,8 +213,8 @@ def process_subpileup(
             pl.lit(contig_name).alias("contig"),
             pl.lit(modtype).alias("mod_type")
         ).with_columns([
-            pl.col("model").map_elements(lambda x: x._alpha).alias("alpha"),
-            pl.col("model").map_elements(lambda x: x._beta).alias("beta")
+            pl.col("model").map_elements(lambda x: x._alpha, return_dtype=pl.Float64).alias("alpha"),
+            pl.col("model").map_elements(lambda x: x._beta, return_dtype=pl.Float64).alias("beta")
         ]).drop("model")
         return identified_motifs
